+++ CHANGELOG.md
@@ -10,7 +10,7 @@ FEATURES:
 * **New Resource:** `panos_general_settings`
 * **New Resource:** `panos_management_profile`
 * **New Resource:** `panos_nat_policy`
-* **New Resource:** `panos_security_policy`
+* **New Resource:** `panos_security_policies`
 * **New Resource:** `panos_service_group`
 * **New Resource:** `panos_service_object`
 * **New Resource:** `panos_virtual_router`